Stop requiring a pystub for register_fake by default
Previously, if someone used `register_fake` to add a fake impl for an
operator defined in C++, we would require them to add a
`m.has_python_registration(<module>)` call to C++. This was to avoid
situations where a user imported the C++ operator without importing the
fake impl.

This "breaks" open registration: there's no way to add a fake impl
outside of a repository that defines an operator, so we want to turn
this behavior off by default in open source.

Test Plan:
- existing tests

ghstack-source-id: 671705beccf3661fed3469a6039eb042ddc0635e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124064